uint32_t --> uint16_t for gpio_base in ich_gpio_offsets struct
git-svn-id: https://mosys.googlecode.com/svn/trunk@458 f72fe4e1-3c20-b1ed-aa54-d76f48875c1e

+++ drivers/intel/ich_gpio.c
@@ -37,7 +37,7 @@ static uint8_t ich_lvl_offsets[] = { 0x0c, 0x38, 0x48 };
 static uint8_t ich_gpio_inv_offset = 0x2c;	/* for port 1 only */
 
 struct {
-	uint32_t gpio_base;
+	uint16_t gpio_base;
 	uint8_t use_sel;
 	uint8_t io_sel;
 	uint8_t lvl;
